Create TodoItem Class
Decided to make a separate module and class for creating a todo item so
that I am following the single responsibility principle, but also to
make things easier to update in the future.

I think making a module/class for 'projects' or separate todo lists
will come next, but I am not sure exactly where to go from there.