Add Project class/module and update todoItem class
Decided to make a project class in addition to the todoItem so that
both could function separately. Currently, the idea is to have all info
of a todo and any methods that change the todo itself to live on that
module. For the project class, I am just naming the project and then
having an empty array with a method to add a todo.

This make sense to me know so we will see.