Add DOM logic to display All Projects
I decided to change the All Projects List from a class to a IIFE as I
thought it would be easier to pull into the DOMstuff class for adding
the functionality to a button event listener. Not sure this is the best
solution, but everything is still separate and easier to follow than
past projects.