Add DOM logic to display all tasks in one project
Decided to make each project a button to click on the View All Projects
page so that all of that single project's tasks render. I added an event
listener to each button to accomplish this and I had to make it an arrow
function due to some 'this' context I don't full understand.