Add DOM logic to add a Project and Task
I had to refactor the code to change the Project Class to include some
static properties and methods as I was having an issue with being able
to have the task be added to the project selected when adding a task. I
think that it is working correctly now so I am committing, but may need
to refactor again once more functionality gets added.