Add DOM logic to edit an existing todoItem
I was able to get the add task modal to include the current values for a
todo when trying to edit it. It was difficult to figure out how to
populate the correct project in the project drop down, but ultimately a
mix of using a option css selector along with an array search to pick
the correct project did the trick.

Next step is to figure out how to have the actual todo be updated once
it is submitted.