Add logic to be able to change the project of a todo
This took me a while to figure out and to be honest, I will need to
review my own code again as it felt like I was googleing and just
throwing some duct tape at my code to make the solution work. However,
it seems to be working now so I will take the win.