Add state to application using localStorage
After reviewing localStorage, I thought it would be easiest to add
another method to the Project Class to update the local storage with
the current array of all Projects any time a new todo or project was
created. This was easy enough.

Then, I tried to retrieve the localStorage and ran into issues anytime
I needed to use a method on one of the instances of the project so the
easiest solution I found was just to take the objects in the
localStorage array and just run them through the Project class on the
intial page load.

It is currently working and I am excited with that, but I think I am
still a little confused or fuzzy on how it all works haha.