Add logic to cancel add todo and project modals
Used a mix of formnovalidate on the cancel button along with
dialog.close() to close the dialog/modal without any errors. While
adding this functionality, I realized that the submit button does not
currently validate as I am using e.preventDefault on the submit button.

In the future, I will need to add a custom validation checker so that
incorrect data is not saved. Not sure I want to add this now as it was
not part of the previous lessons or project points.