Add custom validation to add todo and project dialogs
This was a hacky solution as we haven't learned custom validations in
the course yet so I just used some solution I found on Google. The
reason I couldn't just get away with built in validations is that I was
using e.preventDefault on the dialog/modals and this stops the browser
validation.

Going to relook over the project details to see if I missed anything and
then I think I will pack this one in to continue learning and will
revisit later on with new knowledge.