Add logic to complete a todo
This was the last thing missing from the project requirements so I added
a simple and likely hokey solution of just toggling a strikethrough css
class to make the todo divs have lines through them when clicked.

I am going to complete the project for now and come back to make updates
if needed in the future.